requirements.txt
# MASS (Flowstate) - FastAPI Backend
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Config & validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Caching & sessions
redis>=5.0.0

# Task queue
celery[redis]>=5.3.0

# ML/AI
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.26.0

# API docs (built-in OpenAPI, optional extras)
# Monitoring
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0
factory-boy>=3.3.0
faker>=22.0.0
aiosqlite>=0.19.0

# Dev: formatting & linting
ruff>=0.1.0

# Production
gunicorn>=21.0.0
